ham_digital_modes: fix the Rust 1.98 clippy lints (chunks_exact_to_as_chunks, unnecessary_cast)
Seventeen chunks_exact(N) calls with a constant N become as_chunks::<N>()
(same chunks, same dropped remainder) and one redundant u32 cast goes.
cargo clippy --release --all-targets -- -D warnings is now clean under
the pinned 1.98.1 toolchain; all crate tests pass.

+++ daemons/ham_digital_modes/src/ambe/fec.rs
@@ -124,7 +124,7 @@ pub fn hamming_decode(received: u16) -> (u16, u32) {
     let mut best_data = 0u16;
     let mut best_distance = u32::MAX;
     for data in 0u16..2048 {
-        let distance = (hamming_encode(data) ^ received).count_ones() as u32;
+        let distance = (hamming_encode(data) ^ received).count_ones();
         if distance < best_distance {
             best_distance = distance;
             best_data = data;
